services/ml/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.1.0
    # via tensorboard
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
anyio==4.8.0
    # via starlette
black==25.1.0
    # via detectron2
click==8.1.8
    # via
    #   black
    #   uvicorn
cloudpickle==3.1.1
    # via detectron2
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@main
    # via -r requirements.in
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.4
    # via -r requirements.in
filelock==3.18.0
    # via torch
fonttools==4.56.0
    # via matplotlib
fsspec==2025.3.0
    # via torch
fvcore==0.1.5.post20221221
    # via detectron2
grpcio==1.71.0
    # via tensorboard
h11==0.14.0
    # via uvicorn
hydra-core==1.3.2
    # via detectron2
idna==3.10
    # via anyio
iopath==0.1.9
    # via
    #   detectron2
    #   fvcore
jinja2==3.1.6
    # via torch
kiwisolver==1.4.8
    # via matplotlib
markdown==3.7
    # via tensorboard
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.10.1
    # via
    #   detectron2
    #   pycocotools
mpmath==1.3.0
    # via sympy
mypy-extensions==1.0.0
    # via black
networkx==3.4.2
    # via torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   fvcore
    #   matplotlib
    #   opencv-python
    #   pycocotools
    #   tensorboard
    #   torchvision
omegaconf==2.3.0
    # via
    #   detectron2
    #   hydra-core
opencv-python==4.10.0.84
    # via -r requirements.in
packaging==24.2
    # via
    #   black
    #   detectron2
    #   hydra-core
    #   matplotlib
    #   tensorboard
pathspec==0.12.1
    # via black
pillow==11.0.0
    # via
    #   -r requirements.in
    #   detectron2
    #   fvcore
    #   matplotlib
    #   torchvision
platformdirs==4.3.6
    # via black
portalocker==3.1.1
    # via iopath
protobuf==6.30.1
    # via tensorboard
pycocotools==2.0.8
    # via detectron2
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.6.0
    # via -r requirements.in
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
python-dotenv==1.0.1
    # via pydantic-settings
python-multipart==0.0.17
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   fvcore
    #   omegaconf
    #   yacs
six==1.17.0
    # via
    #   python-dateutil
    #   tensorboard
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
sympy==1.13.3
    # via torch
tabulate==0.9.0
    # via
    #   detectron2
    #   fvcore
tensorboard==2.19.0
    # via detectron2
tensorboard-data-server==0.7.2
    # via tensorboard
termcolor==2.5.0
    # via
    #   detectron2
    #   fvcore
tomli==2.2.1
    # via black
torch==2.2.2
    # via
    #   -r requirements.in
    #   torchaudio
    #   torchvision
torchaudio==2.2.2
    # via -r requirements.in
torchvision==0.17.2
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   detectron2
    #   fvcore
    #   iopath
typing-extensions==4.12.2
    # via
    #   anyio
    #   black
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   torch
    #   uvicorn
uvicorn==0.32.0
    # via -r requirements.in
werkzeug==3.1.3
    # via tensorboard
yacs==0.1.8
    # via
    #   detectron2
    #   fvcore

# The following packages are considered to be unsafe in a requirements file:
# setuptools
